+++ zebra-network/src/peer/load_tracked_client.rs
@@ -2,7 +2,7 @@
 //! reported protocol version.
 
 use std::{
-    net::IpAddr,
+    net::{IpAddr, SocketAddr},
     sync::Arc,
     task::{Context, Poll},
 };
@@ -15,7 +15,7 @@ use tower::{
 use crate::{
     constants::{EWMA_DECAY_TIME_NANOS, EWMA_DEFAULT_RTT},
     peer::{Client, ConnectedAddr, ConnectionInfo},
-    protocol::external::types::Version,
+    protocol::external::{canonical_socket_addr, types::Version},
 };
 
 /// A client service wrapper that keeps track of its load.
@@ -57,10 +57,12 @@ impl LoadTrackedClient {
 
     /// Returns true if this peer connected directly to us from `ip`.
     pub fn is_inbound_direct_from_ip(&self, ip: &IpAddr) -> bool {
+        let expected_ip = canonical_socket_addr(SocketAddr::new(*ip, 0)).ip();
+
         matches!(
             self.connection_info.connected_addr,
             ConnectedAddr::InboundDirect { addr }
-                if addr.remove_socket_addr_privacy().ip() == *ip
+                if canonical_socket_addr(addr.remove_socket_addr_privacy()).ip() == expected_ip
         )
     }
 }